back-end/requirement.txt
certifi==2019.9.11
Django==2.2.5
django-cors-headers==3.1.0
django-extensions==2.2.1
django-jet==1.0.8
djangorestframework==3.10.3
psycopg2==2.8.3
pytz==2019.2
sentry-sdk==0.12.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
raven==6.10.0
prometheus-client==0.7.1
